+++ src/main/java/com/nbu/logistics/services/MyUserDetailsService.java
@@ -1,14 +1,10 @@
 package com.nbu.logistics.services;
 
-import java.util.Collection;
-
 import com.nbu.logistics.config.MyUserPrincipal;
 import com.nbu.logistics.entities.User;
 import com.nbu.logistics.repositories.UsersRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
